Update the UI for catergory

+++ src/app/modules/admin/pages/content/category/list-category/list-category.component.ts
@@ -8,13 +8,14 @@ import {
     OnInit,
     ViewChild,
     ViewEncapsulation,
-    Pipe, PipeTransform
+    Pipe, PipeTransform,
+    inject
 } from '@angular/core';
 import { MatIconModule } from '@angular/material/icon';
 import { ActivatedRoute, Router, RouterLink, RouterOutlet, RouterModule } from '@angular/router';
 
 import { DateTime } from 'luxon';
-import { Observable, Subject, takeUntil } from 'rxjs';
+import { Observable, Subject, switchMap, takeUntil } from 'rxjs';
 import { ContentService } from '../../content.service';
 import { Board } from '../../content.models';
 import { MatDrawer, MatSidenavModule } from '@angular/material/sidenav';
@@ -22,11 +23,18 @@ import { FuseMediaWatcherService } from '@fuse/services/media-watcher';
 import { FuseCardComponent } from '@fuse/components/card';
 import { MatButtonModule } from '@angular/material/button';
 import { Category, ModelContentCategoryDelete } from 'app/shared/api/model/models';
-import { ContentCategoryService } from 'app/shared/api/services/api';
-import { AsyncPipe, CommonModule } from '@angular/common';
+import { ContentCategoryService, SharedService } from 'app/shared/api/services/api';
+import { AsyncPipe, CommonModule, DatePipe } from '@angular/common';
 import { MatTooltipModule } from '@angular/material/tooltip';
 import { FuseAlertType } from '@fuse/components/alert/alert.types';
 import { FuseAlertComponent } from '@fuse/components/alert';
+import { MatDialog, MatDialogModule } from '@angular/material/dialog';
+import { ConfirmDialogComponent } from 'app/shared/component/confirm-dialog/confirm-dialog.component';
+import { ConfirmDialog } from 'app/shared/types';
+import { MatMenuModule } from '@angular/material/menu';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { FormsModule, ReactiveFormsModule, UntypedFormControl } from '@angular/forms';
+import { MatInputModule } from '@angular/material/input';
 
 
 @Component({
@@ -45,12 +53,20 @@ import { FuseAlertComponent } from '@fuse/components/alert';
         CommonModule,
         MatTooltipModule,
         FuseAlertComponent,
-        RouterModule
+        RouterModule,
+        MatDialogModule,
+        MatMenuModule,
+        MatFormFieldModule,
+        FormsModule,
+        ReactiveFormsModule,
+        MatInputModule,
+        DatePipe 
     ],
 })
 export class CategoryListComponent implements OnInit, OnDestroy {
     @ViewChild('matCategoryDrawer', { static: true }) matCategoryDrawer: MatDrawer;
-
+    readonly dialog = inject(MatDialog);
+    readonly sharedService = inject(SharedService);
     boards: Board[];
     categories:Category[];
     categories$:Observable<Category[]>;
@@ -64,6 +80,8 @@ export class CategoryListComponent implements OnInit, OnDestroy {
         };
     
     showAlert = false;
+   searchInputControl: UntypedFormControl = new UntypedFormControl();
+   categoriesCount = 0;
     /**
      * Constructor
      */
@@ -114,7 +132,8 @@ export class CategoryListComponent implements OnInit, OnDestroy {
           this.categories$ = this._contentCategoryService.categories$;
                 this._contentCategoryService.categories$
                     .pipe(takeUntil(this._unsubscribeAll))
-                    .subscribe((contacts: Category[]) => {
+                    .subscribe((categories: Category[]) => {
+                        this.categoriesCount = categories.length;
                         // Mark for check
                         this._changeDetectorRef.markForCheck();
                     });
@@ -128,6 +147,17 @@ export class CategoryListComponent implements OnInit, OnDestroy {
                 this._changeDetectorRef.markForCheck();
             }
         });
+
+          // Subscribe to search input field value changes
+                this.searchInputControl.valueChanges
+                    .pipe(
+                        takeUntil(this._unsubscribeAll),
+                        switchMap((query) =>
+                            // Search
+                            this._contentCategoryService.searchCategoriesContentCategorySearchGet(query)
+                        )
+                    )
+                    .subscribe();
     }
 
     updateList() {
@@ -191,29 +221,41 @@ export class CategoryListComponent implements OnInit, OnDestroy {
      }  
     }
 
-    removeCategory(id:string) {
-       const deleteCategoryModel = <ModelContentCategoryDelete>{id:id};
-        this._contentCategoryService.deleteCategoryContentCategoryDeletePost(deleteCategoryModel).subscribe({
-            next:(response) => {
-                if (response.status.toLocaleLowerCase() === "success") {
-                    // Set the alert
-                    this.alert = {
-                        type: 'success',
-                        message: `Category has been removed successfuly.`,
-                    };
-                       // Show the alert
-                       this.showAlert = true;
-                       // Mark for check
-                       this.hideAlert();
-                       this.updateList();
-                }
-            },
-            error:(error)=> {
-                this.showError(error);
+    removeCategory(id: string) {
+        this.sharedService.confirmDialog = <ConfirmDialog>{
+            cancelButtonLabel: "Cancel",
+            confirmButtonLabel: "Delete",
+            message: "Do you really want to delete?",
+            title: "Are you sure?"
+        };
+        const dialogRef = this.dialog.open(ConfirmDialogComponent);
+        dialogRef.afterClosed().subscribe((result: boolean) => {
+            if (result) {
+                const deleteCategoryModel = <ModelContentCategoryDelete>{ id: id };
+                this._contentCategoryService.deleteCategoryContentCategoryDeletePost(deleteCategoryModel).subscribe({
+                    next: (response) => {
+                        if (response.status.toLocaleLowerCase() === "success") {
+                            // Set the alert
+                            this.alert = {
+                                type: 'success',
+                                message: `Category has been removed successfuly.`,
+                            };
+                            // Show the alert
+                            this.showAlert = true;
+                            // Mark for check
+                            this.hideAlert();
+                            this.updateList();
+                        }
+                    },
+                    error: (error) => {
+                        this.showError(error);
+                    }
+                });
             }
         })
     }
 
+    
     /**
      * append unique value against url to reflect image on UI after update.
      * @param url
